# backend/openapi.json
{
  "components": {
    "schemas": {
      "AccessListAuthCreate": {
        "description": "A basic-auth credential to add to an access list (password write-only).",
        "properties": {
          "password": {
            "description": "Plaintext; stored hashed, never returned",
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "AccessListAuthCreate",
        "type": "object"
      },
      "AccessListAuthRead": {
        "description": "Public representation of a basic-auth user (no credential material).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "created_at",
          "updated_at"
        ],
        "title": "AccessListAuthRead",
        "type": "object"
      },
      "AccessListAuthUpdate": {
        "description": "Reset a basic-auth user's password.",
        "properties": {
          "password": {
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "AccessListAuthUpdate",
        "type": "object"
      },
      "AccessListClientCreate": {
        "description": "An allow/deny rule for an IP address, CIDR range, or ``all``.",
        "properties": {
          "address": {
            "description": "IP, CIDR, or 'all'",
            "maxLength": 255,
            "minLength": 1,
            "title": "Address",
            "type": "string"
          },
          "directive": {
            "$ref": "#/components/schemas/AccessListDirective",
            "description": "allow or deny"
          }
        },
        "required": [
          "address",
          "directive"
        ],
        "title": "AccessListClientCreate",
        "type": "object"
      },
      "AccessListClientRead": {
        "description": "Public representation of an IP client rule.",
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "directive": {
            "$ref": "#/components/schemas/AccessListDirective"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "address",
          "directive",
          "created_at",
          "updated_at"
        ],
        "title": "AccessListClientRead",
        "type": "object"
      },
      "AccessListClientUpdate": {
        "description": "Partial update of a client rule.",
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "directive": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccessListDirective"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AccessListClientUpdate",
        "type": "object"
      },
      "AccessListCreate": {
        "description": "Payload to create an access list, optionally seeding users and rules inline.",
        "properties": {
          "auth_users": {
            "items": {
              "$ref": "#/components/schemas/AccessListAuthCreate"
            },
            "title": "Auth Users",
            "type": "array"
          },
          "clients": {
            "items": {
              "$ref": "#/components/schemas/AccessListClientCreate"
            },
            "title": "Clients",
            "type": "array"
          },
          "name": {
            "description": "Human-readable name",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pass_auth": {
            "default": false,
            "description": "Forward the Authorization header to the upstream",
            "title": "Pass Auth",
            "type": "boolean"
          },
          "satisfy_any": {
            "default": false,
            "description": "Pass if EITHER gate (auth OR ip) is satisfied; false requires both",
            "title": "Satisfy Any",
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ],
        "title": "AccessListCreate",
        "type": "object"
      },
      "AccessListDirective": {
        "description": "Client-rule directive for an access list entry.",
        "enum": [
          "allow",
          "deny"
        ],
        "title": "AccessListDirective",
        "type": "string"
      },
      "AccessListRead": {
        "description": "Public representation of an access list, including users and rules.",
        "properties": {
          "auth_users": {
            "items": {
              "$ref": "#/components/schemas/AccessListAuthRead"
            },
            "title": "Auth Users",
            "type": "array"
          },
          "client_rules": {
            "items": {
              "$ref": "#/components/schemas/AccessListClientRead"
            },
            "title": "Client Rules",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "description": "Human-readable name",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pass_auth": {
            "default": false,
            "description": "Forward the Authorization header to the upstream",
            "title": "Pass Auth",
            "type": "boolean"
          },
          "satisfy_any": {
            "default": false,
            "description": "Pass if EITHER gate (auth OR ip) is satisfied; false requires both",
            "title": "Satisfy Any",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "AccessListRead",
        "type": "object"
      },
      "AccessListUpdate": {
        "description": "Partial update of an access list's own attributes (not its users/rules).",
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "pass_auth": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Auth"
          },
          "satisfy_any": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Satisfy Any"
          }
        },
        "title": "AccessListUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "MegooPM",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/access-lists": {
      "get": {
        "description": "List all access lists with their users and rules. Admin-only.",
        "operationId": "list_access_lists_api_v1_access_lists_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccessListRead"
                  },
                  "title": "Response List Access Lists Api V1 Access Lists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Access Lists",
        "tags": [
          "access-lists"
        ]
      },
      "post": {
        "description": "Create an access list, optionally seeding users and rules inline. Admin-only.",
        "operationId": "create_access_list_api_v1_access_lists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessListCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessListRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Access List",
        "tags": [
          "access-lists"
        ]
      }
    },
    "/api/v1/access-lists/{access_list_id}": {
      "delete": {
        "description": "Delete an access list. Attached hosts are detached (FK SET NULL).",
        "operationId": "delete_access_list_api_v1_access_lists__access_list_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "access_list_id",
            "required": true,
            "schema": {
              "title": "Access List Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Access List",
        "tags": [
          "access-lists"
        ]
      },
      "get": {
        "description": "Fetch a single access list. Admin-only.",
        "operationId": "get_access_list_api_v1_access_lists__access_list_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "access_list_id",
            "required": true,
            "schema": {
              "title": "Access List Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessListRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Access List",
        "tags": [
          "access-lists"
        ]
      },
      "patch": {
        "description": "Update an access list's own attributes (name/satisfy_any/pass_auth).",
        "operationId": "update_access_list_api_v1_access_lists__access_list_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "access_list_id",
            "required": true,
            "schema": {
              "title": "Access List Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessListUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessListRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Access List",
        "tags": [
          "access-lists"
        ]
      }
    },
    "/api/v1/access-lists/{access_list_id}/auth-users": {
      "post": {
        "description": "Add a basic-auth user to an access list. Admin-only.",
        "operationId": "add_auth_user_api_v1_access_lists__access_list_id__auth_users_post",
        "parameters": [
          {
            "in": "path",
            "name": "access_list_id",
            "required": true,
            "schema": {
              "title": "Access List Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessListAuthCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessListAuthRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add Auth User",
        "tags": [
          "access-lists"
        ]
      }
    },
    "/api/v1/access-lists/{access_list_id}/auth-users/{user_id}": {
      "delete": {
        "description": "Remove a basic-auth user from an access list. Admin-only.",
        "operationId": "remove_auth_user_api_v1_access_lists__access_list_id__auth_users__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "access_list_id",
            "required": true,
            "schema": {
              "title": "Access List Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Remove Auth User",
        "tags": [
          "access-lists"
        ]
      },
      "patch": {
        "description": "Reset a basic-auth user's password. Admin-only.",
        "operationId": "set_auth_password_api_v1_access_lists__access_list_id__auth_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "access_list_id",
            "required": true,
            "schema": {
              "title": "Access List Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessListAuthUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessListAuthRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Set Auth Password",
        "tags": [
          "access-lists"
        ]
      }
    },
    "/api/v1/access-lists/{access_list_id}/clients": {
      "post": {
        "description": "Add an allow/deny client rule to an access list. Admin-only.",
        "operationId": "add_client_rule_api_v1_access_lists__access_list_id__clients_post",
        "parameters": [
          {
            "in": "path",
            "name": "access_list_id",
            "required": true,
            "schema": {
              "title": "Access List Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessListClientCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessListClientRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add Client Rule",
        "tags": [
          "access-lists"
        ]
      }
    },
    "/api/v1/access-lists/{access_list_id}/clients/{rule_id}": {
      "delete": {
        "description": "Remove a client rule from an access list. Admin-only.",
        "operationId": "remove_client_rule_api_v1_access_lists__access_list_id__clients__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "access_list_id",
            "required": true,
            "schema": {
              "title": "Access List Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Remove Client Rule",
        "tags": [
          "access-lists"
        ]
      },
      "patch": {
        "description": "Update a client rule within an access list. Admin-only.",
        "operationId": "update_client_rule_api_v1_access_lists__access_list_id__clients__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "access_list_id",
            "required": true,
            "schema": {
              "title": "Access List Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessListClientUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessListClientRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Client Rule",
        "tags": [
          "access-lists"
        ]
      }
    },
...
  }
}